pyproject.toml
[project]
name = "chattix"
version = "0.1.0"
description = "Real-time chat with Litestar, WebSockets, Redis, PostgreSQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litestar[standard]>=2.10",
    "multipart>=1.2",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "redis[hiredis]>=5.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "pydantic-settings>=2.2",
    "email-validator>=2.0",
]

[project.scripts]
chattix = "chattix.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["chattix"]

[tool.ruff]
line-length = 100
target-version = "py311"
